Moved context/action dictionary into the ConnectionManager
